fix(data-block): move collection hover actions next to vote buttons in list and gallery
Extends the previous alignment fix to COLLECTION-source data blocks. The popover (...) menu and side panel button used to render at the top-right of the title area inside CollectionMetadata, taking a 56px right padding and leaving the icons vertically offset from the vote buttons. Extract the hover-action row into a new CollectionRowActions component and render it inline next to EntityVoteButtons in list and gallery views, freeing up horizontal space for longer titles and putting the icons on a single baseline with the vote arrows.

CollectionMetadata still renders the same actions inline by default (preserving the existing layout for table and bulleted-list views). A new hideHoverActions prop opts the gallery/list paths out of the inline rendering. The verified checkmark continues to sit next to the title, with a tighter pr-6 reservation when no hover actions need pr-14.
