+++ apps/web/partials/blocks/table/table-block-gallery-item.tsx
@@ -19,6 +19,7 @@ import { SelectEntity } from '~/design-system/select-entity';
 
 import type { onChangeEntryFn, onLinkEntryFn } from '~/partials/blocks/table/change-entry';
 import { CollectionMetadata } from '~/partials/blocks/table/collection-metadata';
+import { CollectionRowActions } from '~/partials/blocks/table/collection-row-actions';
 import { DataBlockOpenSidePanelButton } from '~/partials/blocks/table/data-block-open-side-panel-button';
 import { EditModeNameField } from '~/partials/blocks/table/edit-mode-name-field';
 import { EntityVoteButtons } from '~/partials/entity-page/entity-vote-buttons';
@@ -273,7 +274,7 @@ export function TableBlockGalleryItem({
                 relationId={relationId}
                 verified={verified}
                 onLinkEntry={onLinkEntry}
-                showSidePanel={!isPlaceholder}
+                hideHoverActions
                 openedWithMainViewEditing={isEditing}
               >
                 <Link entityId={rowEntityId} spaceId={currentSpaceId} href={href}>
@@ -283,13 +284,26 @@ export function TableBlockGalleryItem({
             )}
           </div>
           <div className="flex shrink-0 items-center gap-1">
-            {source.type !== 'COLLECTION' && !isPlaceholder && (
+            {!isPlaceholder && (
               <div className="opacity-0 transition duration-200 group-hover:opacity-100">
-                <DataBlockOpenSidePanelButton
-                  entityId={rowEntityId}
-                  entitySpaceId={nameCell?.space ?? currentSpaceId}
-                  openedWithMainViewEditing={isEditing}
-                />
+                {source.type === 'COLLECTION' ? (
+                  <CollectionRowActions
+                    isEditing={false}
+                    currentSpaceId={currentSpaceId}
+                    entityId={rowEntityId}
+                    spaceId={nameCell?.space}
+                    relationId={relationId}
+                    verified={verified}
+                    onLinkEntry={onLinkEntry}
+                    openedWithMainViewEditing={isEditing}
+                  />
+                ) : (
+                  <DataBlockOpenSidePanelButton
+                    entityId={rowEntityId}
+                    entitySpaceId={nameCell?.space ?? currentSpaceId}
+                    openedWithMainViewEditing={isEditing}
+                  />
+                )}
               </div>
             )}
             <EntityVoteButtons entityId={rowEntityId} spaceId={currentSpaceId} />
